+++ demo/src/main/java/com/orange/ods/demo/ui/components/VariantDetailScreen.kt
@@ -14,6 +14,8 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import com.google.accompanist.pager.ExperimentalPagerApi
+import com.orange.ods.demo.ui.TopAppBarConfiguration
+import com.orange.ods.demo.ui.components.appbars.top.ComponentTopAppBar
 import com.orange.ods.demo.ui.components.buttons.ComponentButtons
 import com.orange.ods.demo.ui.components.cards.ComponentCard
 import com.orange.ods.demo.ui.components.chips.VariantChip
@@ -29,6 +31,7 @@ import com.orange.ods.demo.ui.components.textfields.ComponentTextField
 fun VariantDetailScreen(
     variantId: Long,
     updateTopBarTitle: (Int) -> Unit,
+    updateTopAppBar: (TopAppBarConfiguration) -> Unit,
     updateTopAppBarTabs: (TabsConfiguration) -> Unit
 ) {
     val component = remember { components.firstOrNull { component -> component.variants.any { variant -> variant.id == variantId } } }
@@ -37,6 +40,7 @@ fun VariantDetailScreen(
     variant?.let {
         updateTopBarTitle(variant.titleRes)
         when (component) {
+            Component.AppBarsTop -> ComponentTopAppBar(updateTopAppBar)
             Component.Buttons -> ComponentButtons(variant = variant)
             Component.Cards -> ComponentCard(variant = variant)
             Component.Chips -> if (variant == Variant.ChipFilter) VariantChipFilter() else VariantChip()